+++ src/main/java/ru/yandex/practikum/filmorate/storage/GenreStorage.java
@@ -15,9 +15,4 @@ public interface GenreStorage {
      * получить все жанры
      */
     List<Genre> getAllGenres();
-
-    /**
-     * список id жанров фильма по его id
-     */
-    List<Integer> getGenreIdsByFilmId(long id);
 }
