--- src/main/java/ru/yandex/practikum/filmorate/storage/IdGenerator.java
@@ -1,15 +1,0 @@
-package ru.yandex.practikum.filmorate.storage;
-
-import lombok.Setter;
-
-@Setter
-public class IdGenerator {
-    private long id = 1L;
-
-    public IdGenerator() {
-    }
-
-    public long getId() {
-        return id++;
-    }
-}
